requirements.txt
requests
beautifulsoup4
fpdf
mysql-connector-python

